moved the Automata2D builders from the ConsoleRunHelper class to the Automata2D class
